+++ src/iv.jl
@@ -2,6 +2,8 @@ module IV
 
 export iv
 
+import ..ErrorsInVariables: Either, Left, Right
+
 
 """
     iv(X::Matrix, y::Vector, Z::Matrix)::Vector
@@ -22,20 +24,22 @@ computed using the formula:
     betas = inv(part * X) * part * y
 ```
 """
-function iv(X::Matrix, y::Vector, Z::Matrix)::Vector 
+function iv(X::Matrix, y::Vector, Z::Matrix)::Either
 
     numberofinstruments = size(Z, 2)
     
     numberofpredictors = size(X, 2)
 
-    @assert numberofinstruments >= numberofpredictors 
+    if numberofinstruments < numberofpredictors 
+        return Left("Error: Number of instruments must be greater than or equal to the number of predictors.")
+    end
 
     if numberofinstruments == numberofpredictors
-        return inv(Z'X)Z'y
+        return Right(inv(Z'X)Z'y)
     else
         part = X' * Z * inv(Z' * Z) * Z' 
         betas = inv(part * X) * part * y
-        return betas
+        return Right{Vector}(betas)
     end
 
 end